fixed the icons and board func
the icons now appear in the menus.
also, the delete board and add board were causing crashes on existing and non existing boards - respectively.
fixed that too, if a board for a user already exists with the same name, wont create and a message would pop.
same goes for deletion - if a board doesn't exist, you can't delete it and thus a message would pop.