fixed id bug
appearently, after creating two boards and deleting the first of the two, the boardfacade
would hold the same id as the second board. upon creation of another board, the id cannot
be used and an exception is thrown. fixed it by updating the boardID, after board deletion
to be the largest ID in the db + 1.